requirements.txt
twilio
pulsar-client
python-dotenv
cassandra-driver